+++ retro-go/components/gifdec/gifdec.h
@@ -41,7 +41,7 @@ typedef struct gd_GIF {
 gd_GIF *gd_open_gif(const char *fname);
 int gd_get_frame(gd_GIF *gif);
 void gd_render_frame(gd_GIF *gif, uint8_t *buffer);
-void gd_render_frame565(gd_GIF *gif, uint8_t *buffer);
+void gd_render_frame565(gd_GIF *gif, uint16_t *buffer);
 void gd_rewind(gd_GIF *gif);
 void gd_close_gif(gd_GIF *gif);
